+++ README.md
@@ -37,4 +37,6 @@
 - [ ] RequestBody 를 처리할 수 있는 ArgumentResolver 추가
   - [x] `@RequestBody` 애너테이션 추가
   - [ ] RequestParameterUtils를 활용한 query string, json 지원
- 
+- [x] AbstractModelArgumentResolver 추상화 
+  - [ ] RequestBodyArgumentResolver에서 동일한 코드를 재사용 
+
